+++ src/audio_visual/accel/src/tools/level_parser.c
@@ -6,7 +6,7 @@ static FileContents read_file_contents(const char* const file_name) {
 	if (file == NULL) FAIL(OpenFile, "could not open a file with the path of '%s'.", file_name);
 
 	fseek(file, 0l, SEEK_END); // Set file position to end
-	const long num_bytes = ftell(file);
+	const size_t num_bytes = (size_t) ftell(file);
 	fseek(file, 0l, SEEK_SET); // Rewind file position
 
 	char* const data = malloc(num_bytes + 1l);
